refactor: Rename the COCO directory
