+++ convert/src/coco_format/exceptions.py
