+++ convert/src/coco_format/utils/render_pdf_page.py
