+++ convert/src/routers/coco.py
@@ -8,15 +8,15 @@ from requests import HTTPError
 from tenant_dependency import TenantData, get_tenant_info
 
 from src.config import minio_client, settings
-from src.legacy.coco_export.convert import ConvertToCoco, ExportBadgerdoc
-from src.legacy.coco_export.export_service import (
+from src.coco_format.coco_export.convert import ConvertToCoco, ExportBadgerdoc
+from src.coco_format.coco_export.export_service import (
     export_run,
     export_run_and_return_url,
 )
-from src.legacy.coco_import.convert import ConvertToBadgerdoc
-from src.legacy.coco_import.import_job import create_import_job
-from src.legacy.models import coco
-from src.legacy.utils.s3_utils import get_bucket_path
+from src.coco_format.coco_import.convert import ConvertToBadgerdoc
+from src.coco_format.coco_import.import_job import create_import_job
+from src.coco_format.models import coco
+from src.coco_format.utils.s3_utils import get_bucket_path
 from src.logger import get_logger
 
 router = APIRouter(prefix="/coco", tags=["coco"])
